requirements.txt
tensorflow
numpy
matplotlib
pillow
google_images_search